requirements.txt
# Core web framework
fastapi>=0.104.0
uvicorn>=0.24.0

# LlamaIndex ecosystem (let pip resolve compatible versions)
llama-index
llama-index-llms-openai
llama-index-vector-stores-chroma

# Vector database
chromadb>=0.4.22

# MongoDB
pymongo>=4.6.0
motor>=3.3.0
dnspython>=2.4.0

# AI/ML
openai>=1.3.0
pydantic>=2.5.0

# Utilities
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0
httpx>=0.25.0

# Security
pytz>=2023.3
